Support SAML SSO with omniauth-saml (#11)
This is a batch of updates to make authenticating when SAML SSO, using
the omniauth-saml gem, work correctly. Two main changes:

* Do not include "extra" info in session to avoid infinite as_json loop
* Disable CSRF protection for "create session" endpoint, which is more
  like an API endpoint in a SAML auth flow

**Extra info**

The omniauth-saml provider includes a complex object in the
`["extra"]["response_object"]` key that can't be serialized as JSON. If
you try, it triggers an infinite loop.

This is definitely tight coupling between gems, but...oh well. This
library is small enough that I'm considering just re-implementing it
in our backend. It's more of a hassle to manage a separate package and
try to pretend we aren't modifying it to fit our specific use case.

**Disabling CSRF**

See the comment in the code for a justification of this change, which
might otherwise look like an obvious security vulnerability.

**Test Plan**

Tested in conjunction with https://github.com/AtoB-Developers/backend/pull/2739